add tests for new workspace cases
